chore: TDD 이후 클래스 위치 이동
